+++ frontend/app/src/composables/api/history/index.ts
@@ -1,13 +1,14 @@
 import type { ActionResult } from '@rotki/common';
-import type { AllLocationResponse } from '@/types/location';
 import { api } from '@/services/rotkehlchen-api';
 import { handleResponse, validStatus, validWithSessionStatus } from '@/services/utils';
+import { type AllLocationResponse, type LocationLabel, LocationLabels } from '@/types/location';
 import { ReportProgress } from '@/types/reports';
 
 interface UseHistoryApiReturn {
   getProgress: () => Promise<ReportProgress>;
   fetchAssociatedLocations: () => Promise<string[]>;
   fetchAllLocations: () => Promise<AllLocationResponse>;
+  fetchLocationLabels: () => Promise<LocationLabel[]>;
 }
 
 export function useHistoryApi(): UseHistoryApiReturn {
@@ -35,9 +36,18 @@ export function useHistoryApi(): UseHistoryApiReturn {
     return handleResponse(response);
   };
 
+  const fetchLocationLabels = async (): Promise<LocationLabel[]> => {
+    const response = await api.instance.get<ActionResult<LocationLabel[]>>('/locations/labels', {
+      validateStatus: validStatus,
+    });
+
+    return LocationLabels.parse(handleResponse(response).filter(item => item.locationLabel));
+  };
+
   return {
     fetchAllLocations,
     fetchAssociatedLocations,
+    fetchLocationLabels,
     getProgress,
   };
 }